Reorganised components folder as it was getting to be a bit too big
