+++ ui/src/components/Ratings/EmojiScore.tsx
